Moved client message to its own subproject
