+++ org-code-javabuilder/lib/src/main/java/dev/javabuilder/LocalOutputAdapter.java
@@ -1,8 +1,8 @@
 package dev.javabuilder;
 
 import java.io.PrintStream;
-import org.code.javabuilder.ClientMessage;
 import org.code.javabuilder.OutputAdapter;
+import org.code.util.ClientMessage;
 
 /** Intended for local testing only. Passes output to the provided PrintStream */
 public class LocalOutputAdapter implements OutputAdapter {
