+++ org-code-javabuilder/lib/src/main/java/org/code/javabuilder/OutputAdapter.java
@@ -1,5 +1,7 @@
 package org.code.javabuilder;
 
+import org.code.util.ClientMessage;
+
 public interface OutputAdapter {
   /** @param message An output from the user program */
   void sendMessage(ClientMessage message);
