+++ org-code-javabuilder/lib/src/main/java/org/code/javabuilder/SystemOutMessage.java
@@ -1,5 +1,8 @@
 package org.code.javabuilder;
 
+import org.code.util.ClientMessage;
+import org.code.util.ClientMessageType;
+
 /** A message directed to the client's terminal. Equivalent to System.out.print. */
 public class SystemOutMessage extends ClientMessage {
   public SystemOutMessage(String value) {
