+++ org-code-javabuilder/lib/src/main/java/org/code/javabuilder/UserFacingExceptionMessage.java
@@ -1,6 +1,8 @@
 package org.code.javabuilder;
 
 import java.util.HashMap;
+import org.code.util.ClientMessage;
+import org.code.util.ClientMessageType;
 
 /** An error message directed to the user. Equivalent of a user-visible 500 error. */
 public class UserFacingExceptionMessage extends ClientMessage {
