+++ org-code-javabuilder/lib/src/test/java/org/code/javabuilder/UserFacingExceptionMessageTest.java
@@ -3,6 +3,7 @@ package org.code.javabuilder;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.HashMap;
+import org.code.util.ClientMessage;
 import org.junit.jupiter.api.Test;
 
 public class UserFacingExceptionMessageTest {
