+++ org-code-javabuilder/settings.gradle
@@ -10,3 +10,4 @@
 rootProject.name = 'org-code-javabuilder'
 include('lib')
 include('neighborhood')
+include('util')
