+++ org-code-javabuilder/util/src/main/java/org/code/util/ClientMessage.java
@@ -1,4 +1,4 @@
-package org.code.javabuilder;
+package org.code.util;
 
 import java.util.HashMap;
 import org.json.JSONObject;
